上传下载

+++ src/main/java/com/ssm/pojo/File.java
@@ -4,12 +4,16 @@ import lombok.Data;
 
 import java.io.Serializable;
 
+/**
+ * 文件
+ */
 @Data
 public class File implements Serializable {
+	//文件id
 	private Integer fileid;
-
+	//文件名
 	private String filename;
-
+	//文件夹id
 	private Integer folderid;
 
 	private Integer typeid;
@@ -24,7 +28,7 @@ public class File implements Serializable {
 
 	private String hdfspath;
 
-	private Integer filesize;
+	private Long filesize;
 
 	private String mark;
